Rename alert_dialog.dart
