+++ desktop_photo_search/material/lib/src/widgets/policy_dialog.dart
